Move types to the code that owns the type. Ref #3
